+++ as-relevansi/blocks/search-block/index.asset.php
@@ -6,5 +6,5 @@ return array(
         'wp-block-editor',
         'wp-components',
     ),
-    'version' => '0.1.6',
+    'version' => '0.1.7',
 );